Tutor Servise Update Add Speciality

+++ KnowledgePeak_API.Business/Services/Implements/CLassTimeService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using KnowledgePeak_API.Business.Dtos.ClassTimeDtos;
+using KnowledgePeak_API.Business.Exceptions.ClassTIme;
 using KnowledgePeak_API.Business.Exceptions.Commons;
 using KnowledgePeak_API.Business.Services.Interfaces;
 using KnowledgePeak_API.Core.Entities;
@@ -11,11 +12,13 @@ public class CLassTimeService : IClassTimeService
 {
     readonly IMapper _mapper;
     readonly IClassTimeRepository _repo;
+    readonly IClassScheduleRepository _schedule;
 
-    public CLassTimeService(IMapper mapper, IClassTimeRepository repo)
+    public CLassTimeService(IMapper mapper, IClassTimeRepository repo, IClassScheduleRepository schedule)
     {
         _mapper = mapper;
         _repo = repo;
+        _schedule = schedule;
     }
 
     public async Task CreateAsync(ClassTImeCreateDto dto)
@@ -32,6 +35,8 @@ public class CLassTimeService : IClassTimeService
     {
         if (id <= 0) throw new IdIsNegativeException<ClassTime>();
         var clasTime = await _repo.FIndByIdAsync(id);
+        var schedule = await _schedule.IsExistAsync(s => s.ClassTimeId == id);
+        if (schedule == true) throw new LessonTImeUsedInTheCourseScheduleCannotBeDeletedException();
         if (clasTime == null) throw new NotFoundException<ClassTime>();
         await _repo.DeleteAsync(id);
         await _repo.SaveAsync();
@@ -55,6 +60,9 @@ public class CLassTimeService : IClassTimeService
         if (id <= 0) throw new IdIsNegativeException<ClassTime>();
         var classTime = await _repo.FIndByIdAsync(id);
         if (classTime == null) throw new NotFoundException<ClassTime>();
+
+        var schedule = await _schedule.IsExistAsync(s => s.ClassTimeId == id);
+        if (schedule == true) throw new LessonTImeUsedInTheCourseScheduleCannotBeDeletedException();
 
         var exist = await _repo.IsExistAsync(u => (u.StartTime == dto.StartTime && id != u.Id) &&
         (u.EndTime == dto.EndTime && u.Id != id));
